Move BoardInitial class to board initial module
